pyproject.toml
# rlearn/pyproject.toml

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openstarlab_rlearn"
version = "0.1.20"
description = "openstarlab rlearn modeliing package"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name="Kenjiro Ide", email="ide.kenjirog@g.sp.m.is.nagoya-u.ac.jp" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "ai2-tango=1.3.2",
    "bidict=0.23.1",
    "chardet=5.2.0",
    "datasets=3.2.0",
    "fire=0.5.0",
    "japanize-matplotlib=1.1.3",
    "jsonlines=4.0.0",
    "lightning-lite=1.8.6",
    "matplotlib=3.10.0",
    "matplotsoccer=0.0.8",
    "mlflow=2.5.0",
    "opencv-python>=4.11.0",
    "pyarrow=19.0.0",
    "pydantic=2.10.6",
    "pytorch-lightning=2.0.6",
    "scikit-learn=1.6.1",
    "scipy=1.15.1",
    "torch=2.6.0",
    "torchmetrics=1.0.2",
    "tqdm",
]

license = { text = "Apache License 2.0" }

[tool.setuptools.packages.find]
where = ["."]

[project.urls]
Homepage = "https://github.com/open-starlab/RLearn"